+++ tests/bayes/gmm/benchmark_gmm_estimator.py
@@ -5,8 +5,8 @@ import pytorch_lightning as pl
 import torch
 from pytest_benchmark.fixture import BenchmarkFixture  # type: ignore
 from sklearn.mixture import GaussianMixture as SklearnGaussianMixture  # type: ignore
-from pycave.bayes import GaussianMixture
-from pycave.bayes.core.types import CovarianceType
+from torchgmm.bayes import GaussianMixture
+from torchgmm.bayes.core.types import CovarianceType
 from tests._data.gmm import sample_gmm
