+++ tests/bayes/gmm/test_gmm_model.py
@@ -1,6 +1,6 @@
 # pylint: disable=missing-function-docstring
 from torch import jit
-from pycave.bayes.gmm import GaussianMixtureModel, GaussianMixtureModelConfig
+from torchgmm.bayes.gmm import GaussianMixtureModel, GaussianMixtureModelConfig
 
 
 def test_compile():
