+++ tests/clustering/kmeans/benchmark_kmeans_estimator.py
@@ -5,8 +5,8 @@ import pytorch_lightning as pl
 import torch
 from pytest_benchmark.fixture import BenchmarkFixture  # type: ignore
 from sklearn.cluster import KMeans as SklearnKMeans  # type: ignore
-from pycave.clustering import KMeans
-from pycave.clustering.kmeans.types import KMeansInitStrategy
+from torchgmm.clustering import KMeans
+from torchgmm.clustering.kmeans.types import KMeansInitStrategy
 from tests._data.gmm import sample_gmm
