+++ torchgmm/bayes/markov_chain/model.py
